+++ OpenSourceRecipe.IntegrationTests/Tests/UserTests.cs
@@ -1,16 +1,16 @@
 using System.Net;
 using System.Net.Http.Headers;
 using System.Text;
-using Microsoft.Identity.Client;
 using Newtonsoft.Json;
 using OpenSourceRecipes.Models;
 
 namespace OpenSourceRecipe.IntegrationTests;
 
-public class UserEndpoints(CustomWebApplicationFactory<Program> factory)
-    : IClassFixture<CustomWebApplicationFactory<Program>>
+[Collection("Sequential")]
+public class UserEndpoints
 {
-    private readonly HttpClient _client = factory.CreateClient();
+
+    private readonly HttpClient _client = SharedTestResources.Factory.CreateClient();
 
     [Fact]
     public async Task AuthTestEndpointWithoutToken_ShouldFail()
@@ -704,460 +704,6 @@ public class UserEndpoints(CustomWebApplicationFactory<Program> factory)
         Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
     }
 
-    // RECIPE TESTS
-    [Fact]
-    public async Task CreateRecipeEndpoint_ShouldSucceed()
-    {
-
-        // login user
-        var newUser = new
-        {
-            Username = "seededuser",
-            Password = "password"
-        };
-
-        var loginRequest = new HttpRequestMessage(HttpMethod.Post, "api/login")
-        {
-            Content = new StringContent(JsonConvert.SerializeObject(newUser), Encoding.UTF8, "application/json")
-        };
-
-        var loginResponse = await _client.SendAsync(loginRequest);
-
-        var userDetails = loginResponse.Content.ReadAsStringAsync().Result;
-
-        var user = JsonConvert.DeserializeObject<GetLoggedInUserDto>(userDetails);
-
-        _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", user?.Token);
-
-        //Arrange
-        var newRecipe = new
-        {
-            RecipeTitle = "Test Recipe",
-            TagLine = "Yummy xUnit Test",
-            Difficulty = 1,
-            TimeToPrepare = 1,
-            RecipeMethod = "1. Pass this test",
-            RecipeImg = "https://i.redd.it/iewa8k3fl3d61.jpg",
-            Cuisine = "I don't think we need this field, we have CuisineId",
-            user!.UserId,
-            CuisineId = 1,
-        };
-
-        var request = new HttpRequestMessage(HttpMethod.Post, "api/recipes")
-        {
-            Content = new StringContent(JsonConvert.SerializeObject(newRecipe), Encoding.UTF8, "application/json")
-        };
-
-        //Act
-        var response = await _client.SendAsync(request);
-        var content = await response.Content.ReadAsStreamAsync();
-
-        //Assert
-        Assert.Equal(HttpStatusCode.Created, response.StatusCode);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task CreateRecipeEndpointNoRecipe_ShouldFail()
-    {
-        //Arrange
-        var userLogin = new
-        {
-            Username = "seededuser",
-            Password = "password"
-        };
-
-        var loginRequest = new HttpRequestMessage(HttpMethod.Post, "api/login")
-        {
-            Content = new StringContent(JsonConvert.SerializeObject(userLogin), Encoding.UTF8, "application/json")
-        };
-
-        var loginResponse = await _client.SendAsync(loginRequest);
-
-        var userDetails = loginResponse.Content.ReadAsStringAsync().Result;
-
-        var user = JsonConvert.DeserializeObject<GetLoggedInUserDto>(userDetails);
-
-        _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", user?.Token);
-
-        //Act
-        var request = new HttpRequestMessage(HttpMethod.Post, "api/recipes")
-        {
-            Content = new StringContent("", Encoding.UTF8, "application/json")
-        };
-        var response = await _client.SendAsync(request);
-        var content = await response.Content.ReadAsStreamAsync();
-
-        //Assert
-        Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task GetAllRecipesEndpoint_ShouldSucceed()
-    {
-        // get all recipes
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes");
-        var response = await _client.SendAsync(request);
-        var content = await response.Content.ReadAsStringAsync();
-
-        var recipes = JsonConvert.DeserializeObject<List<GetRecipesDto>>(content);
-
-        foreach (var recipe in recipes!)
-        {
-            Assert.Null(recipe.ForkedFromId);
-            Assert.Null(recipe.OriginalRecipeId);
-        }
-
-        // assert
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task GetAllRecipesByCuisine_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes?cuisineId=1");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllRecipesByCuisine_ShouldReturnAnEmptyList()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes?cuisineId=9999");
-        var response = await _client.SendAsync(request);
-
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count == 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllRecipesByUser_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes?userId=1");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllRecipesByUser_ShouldReturnAnEmptyList()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes?userId=9999");
-        var response = await _client.SendAsync(request);
-
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count == 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipes_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        foreach (var recipe in content!)
-        {
-            Assert.NotNull(recipe.ForkedFromId);
-            Assert.NotNull(recipe.OriginalRecipeId);
-        }
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByCuisine_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?cuisineId=1");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        foreach (var recipe in content!)
-        {
-            Assert.NotNull(recipe.ForkedFromId);
-            Assert.NotNull(recipe.OriginalRecipeId);
-        }
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByCuisine_ShouldReturnAnEmptyList()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?cuisineId=9999");
-        var response = await _client.SendAsync(request);
-
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count == 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByUser_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?userId=1");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        foreach (var recipe in content!)
-        {
-            Assert.NotNull(recipe.ForkedFromId);
-            Assert.NotNull(recipe.OriginalRecipeId);
-            Assert.Equal(1, recipe.UserId);
-        }
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByUser_ShouldReturnAnEmptyList()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?userId=9999");
-        var response = await _client.SendAsync(request);
-
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count == 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByForkedFromId_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?forkedFromId=7");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        foreach (var recipe in content!)
-        {
-            Assert.NotNull(recipe.ForkedFromId);
-            Assert.NotNull(recipe.OriginalRecipeId);
-            Assert.Equal(7, recipe.ForkedFromId);
-        }
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByForkedFromId_ShouldReturnAnEmptyList()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?forkedFromId=9999");
-        var response = await _client.SendAsync(request);
-
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count == 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetAllForkedRecipesByOriginalRecipeId_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/forks?originalRecipeId=7");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRecipesDto>>(contentString);
-
-        foreach (var recipe in content!)
-        {
-            Assert.NotNull(recipe.ForkedFromId);
-            Assert.NotNull(recipe.OriginalRecipeId);
-            Assert.Equal(7, recipe.OriginalRecipeId);
-        }
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count > 0);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetRecipeByIdEndpoint_ShouldSucceed()
-    {
-        // get recipe by id
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/1");
-        var response = await _client.SendAsync(request);
-        var content = await response.Content.ReadAsStreamAsync();
-
-        // assert
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.NotNull(content);
-    }
-
-    [Fact]
-    public async Task GetRecipeByIdEndpointNoRecipe_ShouldFail()
-    {
-        // get recipe by id
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/99");
-        var response = await _client.SendAsync(request);
-
-        // assert
-        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task DeleteRecipeEndpoint_ShouldSucceed()
-    {
-        var userLogin = new
-        {
-            Username = "jellyfish",
-            Password = "password"
-        };
-
-        var loginRequest = new HttpRequestMessage(HttpMethod.Post, "api/login")
-        {
-            Content = new StringContent(JsonConvert.SerializeObject(userLogin), Encoding.UTF8, "application/json")
-        };
-
-        var loginResponse = await _client.SendAsync(loginRequest);
-
-        var userDetails = loginResponse.Content.ReadAsStringAsync().Result;
-
-        var user = JsonConvert.DeserializeObject<GetLoggedInUserDto>(userDetails);
-
-        _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", user?.Token);
-
-        // get recipe by id
-        var request = new HttpRequestMessage(HttpMethod.Delete, "api/recipes/1");
-        var response = await _client.SendAsync(request);
-
-        // assert
-        Assert.Equal(HttpStatusCode.NoContent, response.StatusCode);
-
-        var getRecipeRequest = new HttpRequestMessage(HttpMethod.Get, "api/recipes/1");
-        var getRecipeResponse = await _client.SendAsync(getRecipeRequest);
-
-        Assert.Equal(HttpStatusCode.NotFound, getRecipeResponse.StatusCode);
-    }
-
-    [Fact]
-    public async Task DeleteRecipeEndpointNoRecipe_ShouldFail()
-    {
-        var userLogin = new
-        {
-            Username = "jellyfish",
-            Password = "password"
-        };
-
-        var loginRequest = new HttpRequestMessage(HttpMethod.Post, "api/login")
-        {
-            Content = new StringContent(JsonConvert.SerializeObject(userLogin), Encoding.UTF8, "application/json")
-        };
-
-        var loginResponse = await _client.SendAsync(loginRequest);
-
-        var userDetails = loginResponse.Content.ReadAsStringAsync().Result;
-
-        var user = JsonConvert.DeserializeObject<GetLoggedInUserDto>(userDetails);
-
-        _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", user?.Token);
-
-        // get recipe by id
-        var request = new HttpRequestMessage(HttpMethod.Delete, "api/recipes/99");
-        var response = await _client.SendAsync(request);
-
-        // assert
-        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task DeleteAnotherUserRecipe_ShouldFail()
-    {
-        var userLogin = new
-        {
-            Username = "jellyfish",
-            Password = "password"
-        };
-
-        var loginRequest = new HttpRequestMessage(HttpMethod.Post, "api/login")
-        {
-            Content = new StringContent(JsonConvert.SerializeObject(userLogin), Encoding.UTF8, "application/json")
-        };
-
-        var loginResponse = await _client.SendAsync(loginRequest);
-
-        var userDetails = loginResponse.Content.ReadAsStringAsync().Result;
-
-        var user = JsonConvert.DeserializeObject<GetLoggedInUserDto>(userDetails);
-
-        _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", user?.Token);
-
-        var request = new HttpRequestMessage(HttpMethod.Delete, "api/recipes/9");
-
-        var response = await _client.SendAsync(request);
-
-        Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task GetRelevantRecipes_ShouldSucceed()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/relevant/1");
-        var response = await _client.SendAsync(request);
-        var contentString = await response.Content.ReadAsStringAsync();
-
-        var content = JsonConvert.DeserializeObject<List<GetRelevantRecipesDto>>(contentString);
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.True(content!.Count == 5);
-    }
-
-    [Fact]
-    public async Task GetRelevantRecipes_ShouldReturnAnEmptyList()
-    {
-        var request = new HttpRequestMessage(HttpMethod.Get, "api/recipes/relevant/9999");
-        var response = await _client.SendAsync(request);
-
-        Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-    }
-
     // INGREDIENT TESTS
     [Fact]
     public async Task CreateIngredientEndpoint_ShouldSucceed()
@@ -1344,7 +890,7 @@ public class UserEndpoints(CustomWebApplicationFactory<Program> factory)
         var newComment = new
         {
             Comment = "Test Comment",
-            RecipeId = 1,
+            RecipeId = 13,
             UserId = 1,
             Author = "seededuser"
         };
@@ -1619,7 +1165,7 @@ public class UserEndpoints(CustomWebApplicationFactory<Program> factory)
         // Arrange - login user
         var newUser = new
         {
-            Username = "seededuser",
+            Username = "jellyfish",
             Password = "password"
         };
 
@@ -1638,7 +1184,7 @@ public class UserEndpoints(CustomWebApplicationFactory<Program> factory)
         _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", user?.Token);
 
         // Act - Call the API
-        var request = new HttpRequestMessage(HttpMethod.Patch, $"api/comments/20")
+        var request = new HttpRequestMessage(HttpMethod.Patch, $"api/comments/1")
         {
             Content = new StringContent(JsonConvert.SerializeObject(new { Comment = "" }), Encoding.UTF8, "application/json")
         };
