+++ OpenSourceRecipe.IntegrationTests/WebApplicationFactory.cs
@@ -1,10 +1,17 @@
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 
-namespace OpenSourceRecipe.IntegrationTests;
-
-public class CustomWebApplicationFactory<TStartup> : WebApplicationFactory<TStartup> where TStartup : class
+public class SharedTestResources : IDisposable
 {
+    public static readonly WebApplicationFactory<Program> Factory = new WebApplicationFactory<Program>();
+    public HttpClient Client { get; private set; }
+
+    public SharedTestResources()
+    {
+        Client = Factory.CreateClient();
+    }
+
+    public void Dispose()
+    {
+        Client.Dispose();
+    }
 }
